`fn w_avg_rust`: Restore support for negative strides

+++ src/mc.rs
@@ -681,7 +681,7 @@ unsafe fn avg_rust<BD: BitDepth>(
 
 unsafe fn w_avg_rust<BD: BitDepth>(
     mut dst: *mut BD::Pixel,
-    dst_stride: usize,
+    dst_stride: isize,
     tmp1: &[i16; COMPINTER_LEN],
     tmp2: &[i16; COMPINTER_LEN],
     w: usize,
@@ -704,7 +704,7 @@ unsafe fn w_avg_rust<BD: BitDepth>(
 
         tmp1 = &tmp1[w..];
         tmp2 = &tmp2[w..];
-        dst = dst.offset(dst_stride as isize);
+        dst = dst.offset(dst_stride);
     }
 }
 
@@ -1700,7 +1700,7 @@ unsafe extern "C" fn w_avg_c_erased<BD: BitDepth>(
 ) {
     w_avg_rust(
         dst.cast(),
-        dst_stride as usize,
+        dst_stride,
         tmp1,
         tmp2,
         w as usize,
